Finished Linux and Mac support

+++ src/main/java/io/github/intisy/rsync/Main.java
@@ -3,16 +3,20 @@ package io.github.intisy.rsync;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.util.Arrays;
 
 public class Main {
     private static final boolean logToFile = false;
     public static void main(String[] args) {
-        File logFile = new File(Rclone.getPath(), "latest.log");
+        File rclonePath = Rclone.getPath();
+        File logFile = new File(rclonePath, "latest.log");
+
+        if (!rclonePath.exists() && !rclonePath.mkdirs())
+            throw new RuntimeException("Failed to create directory: " + rclonePath);
+
         try {
             if (logToFile && logFile.exists() && !logFile.delete())
                 throw new RuntimeException("Could not delete log file " + logFile);
-            if (logFile.createNewFile() && logToFile) {
+            if (logToFile && logFile.createNewFile()) {
                 PrintStream systemOut = System.out;
                 PrintStream systemErr = System.err;
                 System.setOut(new PrintStream(logFile) {
